+++ src/model/index.ts
@@ -1,6 +1,6 @@
 export type { Config, RequiredType } from "./config";
 export type { Parents } from "./parents";
-export type { Period } from "./period";
+export type { Period, PeriodEx } from "./period";
 export type { TestResult, TestResults } from "./testResult";
 export type { NewTime, Time, TimeStart, TimeStop } from "./time";
 export type { What } from "./what";