+++ src/model/period.ts
@@ -4,3 +4,7 @@ export type Period = What & {
   start: number;
   stop: number;
 };
+
+export type PeriodEx = Period & {
+  startYear: number;
+};