+++ src/topbar/index.ts
@@ -1,3 +1,3 @@
-export { Topbar } from "./Topbar";
 export { Help } from "./Help";
 export { ScrollToTop } from "./ScrollToTop";
+export { AppYearContext, Topbar, useAppYearContext, useSetYear } from "./Topbar";